test(tokenizer): move fixed tests to test_tokenizer.py
Now that issue #46 is fixed, we don't need to segragate the test cases for the
tokenizer bugs into test_tokenizer_bugs.py anymore, they can live in
test_tokenizer.py instead.
